move code monitor search code out of background package
