+++ enterprise/internal/codemonitors/search_test.go
@@ -1,4 +1,4 @@
-package background
+package codemonitors
 
 import (
 	"context"
